Accept markdown content inside Notifications
