+++ aidants_connect_web/templates/aidants_connect_web/espace_responsable/organisation.html
@@ -37,7 +37,7 @@
             data-notification-url-value="{% url 'notification_mark' notification_id=notification.pk %}"
           >
             <h3 class="fr-alert__title">{{ notification.type_label }}</h3>
-            <p>{{ notification.body }}</p>
+            <p>{{ notification.to_html }}</p>
             {% if notification.must_ack %}
               <button
                 class="fr-btn--close fr-btn"
